Refact: split Logic
